+++ Config/ConfigLoader.php
@@ -24,7 +24,7 @@ declare(strict_types=1);
 namespace Bartacus\Bundle\BartacusBundle\Config;
 
 use Bartacus\Bundle\BartacusBundle\ContentElement\Loader\ContentElementConfigLoader;
-use Bartacus\Bundle\BartacusBundle\Middleware\SymfonyRouteResolver;
+use Bartacus\Bundle\BartacusBundle\Middleware\PrepareContentElementRenderer;
 
 /**
  * Delegating central config loader called on various places within TYPO3
@@ -51,13 +51,14 @@ class ConfigLoader
     {
         return [
             'frontend' => [
-                'bartacus/symfony-route-resolver' => [
-                    'target' => SymfonyRouteResolver::class,
+                'bartacus/prepare-content-element-renderer' => [
+                    'target' => PrepareContentElementRenderer::class,
                     'after' => [
+                        'typo3/cms-frontend/tsfe',
                         'typo3/cms-frontend/site',
-                    ],
-                    'before' => [
-                        'typo3/cms-frontend/base-redirect-resolver',
+                        'typo3/cms-frontend/page-resolver',
+                        'typo3/cms-frontend/page-argument-validator',
+                        'typo3/cms-frontend/prepare-tsfe-rendering',
                     ],
                 ],
             ],